+++ src/app/components/toolbar/toolbar-edit-file.jsx
@@ -6,55 +6,58 @@ import Toolbar from 'material-ui/Toolbar/Toolbar';
 import ToolbarGroup from 'material-ui/Toolbar/ToolbarGroup';
 import ToolbarSeparator from 'material-ui/Toolbar/ToolbarSeparator';
 import ToolbarTitle from 'material-ui/Toolbar/ToolbarTitle';
-import FontIcon from 'material-ui/FontIcon';
 import IconButton from 'material-ui/IconButton';
 import {capimelize} from 'app/libs/helpers';
+import TitleIcon from 'app/components/icon/title-icon';
+import IconList from 'app/components/icon/icon-list';
 
 // APP
 const ToolbarEditFile = class extends React.Component {
 
 	render() {
-		const styles = {
-			toolbar: {
-				justifyContent: 'initial'
-			},
-			button: {
-				marginTop: 5
-			},
-			separator: {
-				margin: '0 5px 0 0px'
-			},
-			icon: {
-				cursor: 'default',
-				margin: '16px 36px 0px -6px',
-				width: 25
-			}
+		const styleSeparator = {
+			margin: '0 5px 0 0px',
+			borderRight: '1px solid white',
+			backgroundColor: 'rgba(0, 0, 0, 0.2)'
 		};
+		const styleIcon = {
+			margin: '4px 22px 0 -18px'
+		};
+		const styleButtons = {marginTop: 5};
+		const styleToolbar = {justifyContent: 'initial'};
 
-		const buttons = [];
-		if (this.props.deleteFileVisible) {
-			buttons.push(
-				<IconButton title="Delete file" className="btDeleteSelected" key={5} style={styles.button} onClick={this.props.onClickDeleteFile}>
-					<FontIcon className="mdi mdi-delete" color="rgba(0, 0, 0, 0.4)" hoverColor="rgba(0, 0, 0, 0.87)"/>
-				</IconButton>
-			);
-		}
+		const dirt = this.props.isDirty;
 
-		if (this.props.saveFileVisible) {
-			buttons.push(
-				<IconButton title="Save file" className="btSaveSelected" key={5} style={styles.button} onClick={this.props.onClickSaveFile}>
-					<FontIcon className="mdi mdi-arrow-down-bold-hexagon-outline" color="rgba(0, 0, 0, 0.4)" hoverColor="rgba(0, 0, 0, 0.87)"/>
-				</IconButton>
-			);
-		}
+		const buttons = [
+			{
+				id: 'save',
+				tooltip: dirt ? 'Save changes' : 'file saved',
+				tooltipPosition: 'top-center',
+				fontIcon: dirt ? 'mdi mdi-content-save' : 'mdi mdi-check',
+				disabled: !dirt
+			}
+		];
+		const iconListProps = {
+			buttons,
+			style: styleButtons,
+			onClick: {
+				save: this.props.onClickSaveFile
+			},
+			iconClassNamePrefix: 'btTest',
+			selectedId: null,
+			raised: true
+		};
+		const renderedButtons = <IconList {...iconListProps}/>;
 
 		return (
-			<Toolbar style={Object.assign(this.props.style || {}, styles.toolbar)}>
-				<FontIcon style={styles.icon} className="mdi mdi-pencil" color="rgba(0, 0, 0, 0.4)" hoverColor="rgba(0, 0, 0, 0.4)"/>
+			<Toolbar style={Object.assign(this.props.style || {}, styleToolbar)}>
+				<IconButton style={styleIcon} title="Back to list" tooltip="Back to list" tooltipPosition="top-center" onClick={this.props.onClickBack}>
+					<TitleIcon className="mdi mdi-arrow-left-bold"/>
+				</IconButton>
 				<ToolbarGroup firstChild lastChild>
-					<ToolbarTitle className={`txt${capimelize(this.props.title)}Title`} text={this.props.title} style={styles.title}/>
-					<ToolbarSeparator style={styles.separator}/>
-					{buttons}
+					<ToolbarTitle className={`txt${capimelize(this.props.title)}Title`} text={this.props.title}/>
+					<ToolbarSeparator style={styleSeparator}/>
+					{renderedButtons}
 				</ToolbarGroup>
 			</Toolbar>
 		);
@@ -68,12 +71,11 @@ ToolbarEditFile.contextTypes = {
 
 ToolbarEditFile.propTypes = {
 	style: React.PropTypes.object,
-	onClickDeleteFile: React.PropTypes.func,
-	deleteFileVisible: React.PropTypes.bool,
 	onClickSaveFile: React.PropTypes.func,
-	saveFileVisible: React.PropTypes.bool,
-	icon: React.PropTypes.string,
-	title: React.PropTypes.string.isRequired
+	title: React.PropTypes.string.isRequired,
+	onClickBack: React.PropTypes.func,
+	isDirty: React.PropTypes.bool,
+	isSaving: React.PropTypes.bool
 };
 
 module.exports = ToolbarEditFile;