+++ CHANGELOG.md
@@ -4,8 +4,12 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a CH
 
 ## [Unreleased]
 
+### Added
+- Add back button to test source editor.
+
 ### Changed
 - Fix style at live machine list.
+- Fix state after saving test source file.
 
 ## [0.7.2]